+++ src/printer.rs
@@ -1,20 +1,20 @@
-use crate::types::{Ast, MalType};
+use crate::types::{Ast, Type};
 
 /// Format an Ast and returns it's string representation
 pub fn pr_str(ast: Ast) -> String {
     pr_type(&ast.root())
 }
 
-/// Recursively format a MalType and returns it's string representation
-fn pr_type(t: &MalType) -> String {
+/// Recursively format a Type and returns it's string representation
+fn pr_type(t: &Type) -> String {
     let mut s = String::new();
     match t {
-        MalType::Nil => s.push_str("nil"),
-        MalType::True => s.push_str("true"),
-        MalType::False => s.push_str("false"),
-        MalType::Integer(integer) => s.push_str(&format!("{}", integer)),
-        MalType::Symbol(symbol) => s.push_str(&format!("{}", symbol)),
-        MalType::List(list) => {
+        Type::Nil => s.push_str("nil"),
+        Type::True => s.push_str("true"),
+        Type::False => s.push_str("false"),
+        Type::Integer(integer) => s.push_str(&format!("{}", integer)),
+        Type::Symbol(symbol) => s.push_str(&format!("{}", symbol)),
+        Type::List(list) => {
             s.push('(');
             s.push_str(
                 &list.iter()