+++ changelog.rst
@@ -6,6 +6,7 @@ CHANGELOG
 - Add initial types (types module) and AST type
 - Add basic lexical and syntax analysis. Read nil, true, false,
   integers, symbols and lists. (reader module)
+- Add printing capabilities (printer module)
 
 0.1.0
 -----